step 5: Introduce HOC for FieldArray
